Move subscription forms, views and templates to separate files
