+++ squarelet/templates/subscriptions/payments.html
@@ -29,6 +29,6 @@
 
 {% block main %}
   <section>
-    {% include "organizations/includes/receipts_card.html" with payments=object_list page_obj=page_obj %}
+    {% include "subscriptions/includes/receipts_card.html" with payments=object_list page_obj=page_obj %}
   </section>
 {% endblock main %}
