+++ squarelet/templates/subscriptions/update_receipt_email.html
