+++ squarelet/templates/subscriptions/update_subscription_frequency.html
